Move ssh related utilities into their own file
Code movement only.
